+++ addon/controllers/i-i-s-caseberry-logging-objects-application-log-l.js
@@ -1,4 +1,11 @@
 /**
- * @module ember-flexberry
+  @module ember-flexberry
+ */
+
+/**
+  Controller for support list form {{#crossLink "IISCaseberryLoggingObjectsApplicationLog"}}IISCaseberryLoggingObjectsApplicationLog{{/crossLink}} record.
+
+  @class IISCaseberryLoggingObjectsApplicationLogLController
+  @extends ListFormController
  */
 export { default } from 'ember-flexberry/controllers/list-form';