+++ tests/dummy/app/transforms/components-examples/flexberry-groupedit/shared/detail-enumeration.js
@@ -1,5 +1,5 @@
 import FlexberryEnumTransormation from 'ember-flexberry/transforms/flexberry-enum';
-import DetailEnumeration from '../../../../enums/components-examples/flexberry-groupedit/settings-example/detail-enumeration';
+import DetailEnumeration from '../../../../enums/components-examples/flexberry-groupedit/shared/detail-enumeration';
 
 export default FlexberryEnumTransormation.extend({
   enum: DetailEnumeration