+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Ember Flexberry Changelog
 ### New Features & improvements
 
+## 2016-06-27
+### Added
+FlexberryGroupeditComponent
+* Add opportunity to observe model changing (it needs to enable flag 'searchForContentChange').
 
 ## 2016-06-24
 ### Breaking changes